renaming:  - ee.pnb.cgitest.pack to ee.pnb.cgitest.archive  - CreateZipService to ZipService
